requirements.txt
asgiref==3.5.2
Django==3.2.9
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.4
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.5
